LayoutTests/platform/chromium/http/tests/security/cross-frame-access-private-browsing-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/cross-frame-iframe.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-private-browsing.html. Domains, protocols and ports must match.

This test checks cross-frame access security checks don't log when private browsing is enabled (rdar://problem/5394877).


Attempting to violate the same-origin policy with private browsing enabled.  If this succeeds the console should not log the violation.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Inner iframe.
